Add missing firewall rules cmr file fixes
